+++ tests/urls.py
@@ -3,7 +3,7 @@
 :license: Apache License 2.0, see LICENSE for more details.
 """
 import nexus
-from django.conf.urls import include, patterns, url
+from django.conf.urls import include, url
 from django.contrib import admin
 
 from gargoyle.compat import subinclude
@@ -16,9 +16,8 @@ def foo(request):
 admin.autodiscover()
 nexus.autodiscover()
 
-urlpatterns = patterns(
-    '',
+urlpatterns = [
     url(r'^nexus/', include(nexus.site.urls)),
     url(r'^admin/', subinclude(admin.site.urls)),
     url(r'^$', foo, name='gargoyle_test_foo'),
-)
+]